Graph fix for progression with intermediate master
Previously, when you add the intermediate master for progression interpolation, the graph (gray lines) was jumped a bit. That's because graph for masters display correct handle vertical position, and for instances the handle vertical position is adjusted. That adjustment was required to correctly draw stem growth curve (gray lines). WIth this update, all intermediate masters also preserves the general curve, so visually it looks more organically for the whole progression.